requirements.txt
absl-py==1.4.0
altgraph==0.17.3
anaconda==0.0.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
beautifulsoup4==4.11.1
blinker==1.8.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.3
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.17.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.4
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.4.8
Jinja2==3.1.4
jiter==0.7.0
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lameenc==1.7.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.0.4
moviepy==1.0.3
mypy==1.11.1
mypy-extensions==1.0.0
numpy==1.23.5
oauthlib==3.2.2
openai==1.54.0
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pdfminer.six==20221105
pikepdf==6.2.8.post1
Pillow==9.4.0
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==4.25.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyserial==3.5
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pywin32-ctypes==0.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.3.2.post1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tqdm==4.66.6
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-requests==2.32.0.20240712
types-Werkzeug==1.0.9
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.14.1
y-py==0.5.5
